Features

+++ server/src/api/roomRoutes.ts
@@ -1,51 +1,63 @@
 import { Request, Response, Router } from "express";
-import { client } from "../config/psql";
+import { pool } from "../config/psql";
 import { user } from "../types/user";
 
 export const router = Router();
 
-router.post("/create", (req: Request, res: Response) => {
-  const roomInfo = req.body;
-  roomInfo.isprivate = false;
-  roomInfo.chatmode = false;
-  roomInfo.creatorid = (req.user as user).userid;
-  const { roomname, roomdesc, isprivate, autospeaker, chatmode, creatorid } =
-    roomInfo;
+router.post("/create", async (req: Request, res: Response) => {
+  try {
+    const roomInfo = req.body;
+    roomInfo.isprivate = false;
+    roomInfo.chatmode = false;
+    roomInfo.creatorid = (req.user as user).userid;
+    const { roomname, roomdesc, isprivate, autospeaker, chatmode, creatorid } =
+      roomInfo;
 
-  console.log(roomInfo);
-  client
-    .query(
+    console.log(roomInfo);
+    const { rows } = await pool.query(
       `
         insert into room (roomname, roomdesc, isprivate, autospeaker, creatorid)
         values ($1, $2, $3, $4, $5) returning roomid
       `,
       [roomname, roomdesc, isprivate, autospeaker, creatorid]
-    )
-    .then(result => {
-      if (result.rows.length > 0) {
-        res.status(200).json(result.rows[0]);
-      }
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+    );
+
+    if (rows.length > 0) {
+      res.status(200).json(rows[0]);
+    } else {
+      res.status(409).json({ msg: "Bad request, couldn't create room" });
+    }
+  } catch (err) {
+    console.log(err);
+    return res.status(500).json(err);
+  }
 });
 
 router.get("/:roomid", async (req: Request, res: Response) => {
   const { roomid } = req.params;
   const { userid } = req.query;
+
+  if (!roomid || !userid) {
+    return res
+      .status(400)
+      .json({ msg: "Bad request, incorrect credentials sent" });
+  }
+
+  const client = await pool.connect();
+
   try {
-    await client.query(
-      `update "user" set currentroomid = $1 where userid = $2`,
-      [roomid, userid]
-    );
-    const { rows: room } = await client.query(
+    await client.query("BEGIN");
+    await pool.query(`update "user" set currentroomid = $1 where userid = $2`, [
+      roomid,
+      userid,
+    ]);
+
+    const { rows: room } = await pool.query(
       `select * from room where roomid = $1`,
       [roomid]
     );
 
-    await client.query(
+    await pool.query(
       `insert into room_permission(roomid, userid, isspeaker, ismod, askedtospeak) 
       values ($1, $2, $3, $4, $5) on conflict do nothing 
      `,
@@ -57,72 +69,107 @@ router.get("/:roomid", async (req: Request, res: Response) => {
         false,
       ]
     );
-    const { rows: participants } = await client.query(
-      `select * from "user"
-     inner join room_permission as rp
-     on rp.userid = "user".userid
-     where rp.roomid = $1
+
+    const { rows: participants } = await pool.query(
+      `select *, 
+      (select count(receiverid) from user_follow where receiverid = "user".userid) as followers, 
+      (select count(causerid) from user_follow where causerid = "user".userid) as following from "user"
+      inner join room_permission as rp
+      on rp.userid = "user".userid
+      where rp.roomid = $1
     `,
       [roomid]
     );
-    res.status(200).json({ ...room[0], participants });
+
+    await client.query("COMMIT");
+    return res.status(200).json({ ...room[0], participants });
   } catch (err) {
-    console.log(err);
-    res.status(500).json(err);
+    await client.query("ROLLBACK");
+    return res.status(500).json(err);
+  } finally {
+    client.release();
   }
 });
 
 router.get("/rooms/live", async (req: Request, res: Response) => {
   try {
-    const { rows: rooms } = await client.query(
-      `select *, array(select username from "user" where currentroomid = room.roomid) as participants from room`
+    const { rows: rooms } = await pool.query(
+      `select *, array(select username from "user" where currentroomid = room.roomid) as participants from room
+       limit 5 
+      `
     );
-    res.status(200).json(rooms);
+
+    return res.status(200).json(rooms);
   } catch (err) {
     console.log(err);
-    res.status(500).json(err);
+    return res.status(500).json(err);
   }
 });
 
 router.get(
-  "/room-permission/:roomId/:userId",
+  "/room-permission/:roomid/:userid",
   async (req: Request, res: Response) => {
+    const { roomid, userid } = req.params;
+
+    if (!roomid || !userid) {
+      return res
+        .status(400)
+        .json({ msg: "Bad request, invalide credentials sent" });
+    }
+
     try {
-      const { roomId, userId } = req.params;
-      const { rows } = await client.query(
+      const { rows } = await pool.query(
         `select u.userid, isspeaker, ismod, askedtospeak, muted
          from room_permission as rp 
          inner join "user" as u
          on rp.userid = u.userid 
          where rp.userid = $1 and roomid = $2`,
-        [userId, roomId]
+        [userid, roomid]
       );
-      res.status(200).json(rows[0]);
+
+      return res.status(200).json(rows[0]);
     } catch (err) {
-      res.status(500).json(err);
       console.log(err);
+      return res.status(500).json(err);
     }
   }
 );
 
 router.post("/leave", async (req: Request, res: Response) => {
-  console.log("leave called");
+  const { userid } = req.user as user;
+  const { roomId: roomid } = req.query;
+
+  console.log("attempting to leave room");
+  console.log("roomid: ", roomid, "userid:", userid);
+  if (!roomid || !userid) {
+    return res
+      .status(400)
+      .json({ msg: "Bad request, invalide credentials sent" });
+  }
+
+  const client = await pool.connect();
+
   try {
-    const { userid } = req.user as user;
-    console.log("[server]: ", req.query, userid);
-    const { roomId } = req.query;
-    await client.query(
+    await client.query("BEGIN");
+
+    await pool.query(
       `update "user" set currentroomid = $1, muted = false where userid = $2`,
       [null, userid]
     );
-    await client.query(
+
+    await pool.query(
       `delete from room_permission where userid = $1 and roomid = $2 `,
-      [userid, roomId]
+      [userid, roomid]
     );
-    res.status(200).json({ msg: "user session cleaned up" });
+
+    await client.query("COMMIT");
+    return res.status(200).json({ msg: "user session cleaned up" });
   } catch (err) {
+    await client.query("ROLLBACK");
     console.log(err);
-    res.status(500).json(err);
+    return res.status(500).json(err);
+  } finally {
+    client.release();
   }
 });
 
@@ -140,17 +187,17 @@ router.put("/room-permission/update", async (req: Request, res: Response) => {
 
     console.log(column, val, roomId, userid);
     if (column === "muted") {
-      await client.query(`update "user" set ${column} = $1 where userid = $2`, [
+      await pool.query(`update "user" set ${column} = $1 where userid = $2`, [
         val,
         userid,
       ]);
     } else {
       val
-        ? await client.query(
+        ? await pool.query(
             `update room_permission set ${column} = $1 where userid = $2 and roomid = $3`,
             [val, userid, roomId]
           )
-        : await client.query(
+        : await pool.query(
             `update room_permission set ${column} = NOT ${column} where userid = $1 and roomid = $2`,
             [userid, roomId]
           );
